updates based on review feedback
