+++ src/index.ts
@@ -2,6 +2,4 @@
 import OktaSignIn from './widget/OktaSignIn';
 
 export default OktaSignIn;
-export { OktaSignIn };
-
 export * from './types';
