+++ src/types/index.ts
@@ -1,5 +1,6 @@
 export * from './api';
 export * from './errors';
+export * from './events';
 export * from './registration';
 export * from './options';
 export * from './results';
